add optional type to Type.ToTypeScriptType for DateTime and DateTimeOffset
This fixes issue #13

As pointed out in the issue, .Net Web API actually sends DateTime and DateTimeOffset back and forth between the front end and the back end as strings. This means that the best practice should be to define these values as strings.

However, Typewriter originally defined these as the JS Date object. So keeping with Typewriter's lead, we are keeping the Date object as the default, but allowing the user to specify a different value for these date types.

This commit adds another argument to ToTypeScriptType that allows the user to specify the Date type.

The unit tests have been updated as well as the documentation.  And this also affects the Parameter.ToTypeScript function as well. It also gets a new argument. (note: at this point in time there are no unit tests for Parameter.ToTypeScript)