Added Action method UrlParameters
UrlParameters returns all the parameters for the action, except the on that is marked as the BodyParameter.
I have added this method because it seems impossible in the Scriban script to be able to remove a Parameter from the IEnumerable list of parameters, because once you start using the array methods, the array becomes a range and is no longer an IEnumerable so you can't do any more parameter work on it.

By providing this Action method, we remove the BodyParam in C# and we retain the type of the array.

For this commit, I added a new method to the Weather controller so that I had an example of an action that did not have a body parameter. I needed this to properly test the case without a body parameter. This addition  of the method leaked into other tests because now there is an extra action to handle.

I also regenerated the documentation so that it picked up the new Action method. The strange thing about that is that it then put all the items in different orders causing a lot more change in this file than should have been.