Added unit tests for ReturnType for .Net Core
Added the different common ways of specifying the return types in .Net Core api controllers:

- just return the type
- return the type wrapped in an ActionResult
- just return the type - async
- return the type wrapped in an ActionResult - async

There is one more way that is possible, but probably not likely. The ReturnType code handles it, but we can't seem to unit test it. This is shown in the ActionWithEnumParam action in the controller. If you uncomment the ProducesResponseType attribute, the return type should be type: WeatherForecast. However, the code in the unit tests does not pick up the attribute argument. I don't know why. It should and it does in a real environment.

For now the unit test is testing that IActionResult is returned, but that shouldn't really be the case.